// src/app/core/interfaces/renglon.ts
export interface Renglones {
    codigo:                  string;
    CodigoAprobacion:        null;
    Descripcion:             string;
    um:                      string;
    material:                number;
    mano:                    number;
    equipos:                 number;
    Peso:                    number;
    Precio:                  number;
    "Precio sin Materiales": number;
    "Horashombre":          number;
    "Horasequipos":         number;
    Tipo:                    string;
    Cemento:                 number;
    Aridos:                  number;
    Asfalto:                 number;
    Carga:                   number;
    Prefabricado:            number;
    "Horasoperario":        number;
    "Horasayudante":        number;
    "Medios Aux":            number;
    "Costo Directo Total":   number;
    "Costos Indirectos":     number;
    "Costo Elaboracion":     number;
    "total":           number;
    Utilidad:                number;
}