// src/app/core/interfaces/apertura.ts
export interface Apertura {
  codigo?: string;
  descripcion?: string;
  tarifaUso?: string;
  tarifaParada?: string;
  Impuesto?: string;
  depreciacion?: string;
  Seguro?: string;
  Combustible?: string;
  aceiteMotor?: string;
  aceiteHidraulico?: string;
  aceiteTransmicion?: string;
  Grasa?: string;
  Electricidad?: string;
  mttoReparacion?: string;
  piezasRepuesto?: string;
  mManoObra?: string;
  neumaticos?: string;
  piezaEspeciales?: string;
  Salario?: string;
  total?: string;
  porCiento?: string;
  otrosGastos?: string;
  Utilidad?: string;
}